+++ CHANGELOG.md
@@ -26,6 +26,7 @@ This project adheres to [Semantic Versioning](https://semver.org).
 	* Change ``ID3v2CommentFrame`` to have only value attribute
 		that contains a single comment.
 	* Change ``ID3v2Frames`` to present a list of comments for ``comment`` key.
+* Rename ``formats.vorbis`` module to ``formats.vorbis_comments``.
 
 ### Removed